pyproject.toml
[tool.poetry]
name = "ozon-api"
version = "0.1.0"
description = ""
authors = ["Владислав Береснев <vladislaw.beresnev2018@yandex.ru>"]
readme = "README.md"
packages = [{include = "ozon_api"}]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "1.4.24"
clickhouse-sqlalchemy = "^0.2.4"
clickhouse-driver = "^0.2.6"
fastapi = "0.73.0"
uvicorn = "^0.22.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"
fastapi-utils = "^0.2.1"
alembic = "1.7.7"
httpx = "^0.24.1"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
